requirements.txt
ultralytics
opencv-python
geocoder
python-telegram-bot
